# Reduce MaxOpenConn to increase performance